bpna converted to tsx and adapted
